# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sport Center API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facilities:
    get:
      operationId: listFacilities
      tags: [facilities]
      summary: List all facilities
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Facility"
    post:
      operationId: createFacility
      tags: [facilities]
      summary: Create a facility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
  /facilities/{id}:
    get:
      operationId: getFacility
      tags: [facilities]
      summary: Get a facility by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
        "404":
          description: Not found
    patch:
      operationId: updateFacility
      tags: [facilities]
      summary: Update a facility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
    delete:
      operationId: deleteFacility
      tags: [facilities]
      summary: Delete a facility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Facility:
      type: object
      required: [id, name, category, pricePerHour, openTime, closeTime, isActive, bookingMode]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        pricePerHour:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        bookingMode:
          type: string
          enum: [time_slot, walk_in]
        minDuration:
          type: integer
        maxDuration:
          type: ["integer", "null"]
        capacity:
          type: ["integer", "null"]
        isActive:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/FacilityImage"
        createdAt:
          type: string
    FacilityImage:
      type: object
      required: [id, facilityId, url, isPrimary]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
    FacilityInput:
      type: object
      required: [name, category, pricePerHour, openTime, closeTime]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        pricePerHour:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        minDuration:
          type: integer
        maxDuration:
          type: integer
        capacity:
          type: integer
        isActive:
          type: boolean
        imageUrls:
          type: array
          items:
            type: string
    FacilityUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        pricePerHour:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        minDuration:
          type: integer
        maxDuration:
          type: integer
        capacity:
          type: integer
        isActive:
          type: boolean
        imageUrls:
          type: array
          items:
            type: string
...